Updated get_file_from_sftp with new kwarg
New kwarg is ignore_host_keys, a boolean that switches whether or
not the host key file should be used when connecting to the server.
The reason for this inclusion is that it's quite unsafe to ignore
host keys, however there are cases where the host key is not
provided.  I also believe there's an active bug in pysftp that
causes an error even when trying to load the known_hosts file so
running without checking (again, in the even that we know the
connection is safe) may fix the issue.

I'll need a better solution in the long term.