Outlook_Mailer has an update() method!
This method allows resetting all values in the mailer.  This way,
after sending a message and clearing the state of the instance,
the instance can be put in a new state by passing the appropriate
variables to the .update() method, which requires the recipient,
subject, and body to be entered.